@@ -4,21 +4,18 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "start": "parcel index.html",
-    "build": "parcel build index.html"
+    "start": "yarn clean-parcel && parcel index.html",
+    "build": "yarn clean-parcel && parcel build index.html",
+    "clean": "yarn clean && rm -rf node_modules",
+    "clean-parcel": "rm -rf .parcel-cache"
   },
   "dependencies": {
-    "react-app-polyfill": "^1.0.0"
-  },
-  "alias": {
-    "react": "../node_modules/react",
-    "react-dom": "../node_modules/react-dom/profiling",
-    "scheduler/tracing": "../node_modules/scheduler/tracing-profiling"
+    "react-app-polyfill": "^3.0.0"
   },
   "devDependencies": {
-    "@types/react": "^16.9.11",
-    "@types/react-dom": "^16.8.4",
-    "parcel": "^1.12.3",
-    "typescript": "^3.4.5"
+    "@types/react": "^18.3.4",
+    "@types/react-dom": "^18.3.0",
+    "parcel": "^2.12.0",
+    "typescript": "^5.5.4"
   }
 }
